fix: co-locate nav data fetch
